Extract AdaptiveMediaSourceEventListener.
